+++ src/payment/models/buy-crypto/services/buy-crypto-notification.service.ts
@@ -1,0 +1,46 @@
+import { Injectable } from '@nestjs/common';
+import { MailService } from 'src/shared/services/mail.service';
+import { IsNull, Not } from 'typeorm';
+import { BuyCryptoRepository } from '../buy-crypto.repository';
+import { BuyCryptoBatchStatus } from '../entities/buy-crypto-batch.entity';
+
+@Injectable()
+export class BuyCryptoNotificationService {
+  constructor(private readonly buyCryptoRepo: BuyCryptoRepository, private readonly mailService: MailService) {}
+
+  async sentNotificationMails(): Promise<void> {
+    const txOutput = await this.buyCryptoRepo.find({
+      where: {
+        recipientMail: IsNull(),
+        mailSendDate: IsNull(),
+        txId: Not(IsNull()),
+        batch: { status: BuyCryptoBatchStatus.COMPLETE },
+      },
+      relations: ['bankTx', 'buy', 'buy.user', 'batch'],
+    });
+
+    for (const tx of txOutput) {
+      await this.mailService.sendTranslatedMail({
+        userData: tx.buy.user.userData,
+        translationKey: 'payment.buyCrypto',
+        params: {
+          buyFiatAmount: tx.amountInEur,
+          // double check the asset
+          buyFiatAsset: 'EUR',
+          buyCryptoAmount: tx.outputAmount,
+          buyCryptoAsset: tx.outputAsset,
+          buyFeePercentage: tx.percentFee,
+          // double check this field
+          buyFeeAmount: tx.absoluteFeeAmount,
+          buyWalletAddress: tx.buy.user.wallet.address,
+          buyTxId: tx.txId,
+        },
+      });
+
+      tx.confirmSentMail();
+
+      // TODO - no need to await? make sure
+      await this.buyCryptoRepo.save(tx);
+    }
+  }
+}
